Convert unicode strings to byte strings
Lxml expects a byte string because XML includes its own encoding
declaration at the beginning of the string [top of the file].

Calling decode() on either a byte string or a unicode string tries
to interpret it as UTF-8 and turn it into a unicode string,
therefore lxml fails to parse with an error "Unicode strings with
encoding declaration are not supported. Please use bytes input or
XML fragments without declaration."

Instead we can just call str() to turn it into a byte string.